requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
PyJWT==1.7.1
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
